chore: fix all clippy deny-lints, add coding guidelines
- Add workspace lint config (13 deny, 1 warn, 4 allow)
- Fix 47 clippy errors across 20 engine files
- Add CONTRIBUTING.md with code style and branch strategy
- Update CLAUDE.md for AI assistant context

+++ pixors-engine/src/composite/mod.rs
@@ -56,8 +56,8 @@ pub fn composite_tile(req: &CompositeRequest<'_>) -> Result<Vec<Rgba<f16>>, Erro
         // Blend each pixel
         for dy in 0..oh {
             for dx in 0..ow {
-                let layer_x = (lx0 as u32) + dx;
-                let layer_y = (ly0 as u32) + dy;
+                let layer_x = lx0 + dx;
+                let layer_y = ly0 + dy;
 
                 let tx = layer_x / req.tile_size;
                 let ty = layer_y / req.tile_size;
